Check import selectors at Typer
Import selectors were checked  at PostTyper before, but that can lead
to confusion when a type error results from a bad import selector.

Fixes #15468
