+++ tests/neg-custom-args/jdk-9-app.scala
@@ -1,4 +1,4 @@
-import java.lang.ProcessHandle
+import java.lang.ProcessHandle // error: not a member
 
 object Jdk9App extends App {
   println(ProcessHandle.current().pid()) // error: not found
