+++ tests/neg/i6056.scala
@@ -1,5 +1,5 @@
 object i0{
-    import i0.i0   // error
+    import i0.i0   // error // error
     def i0={
         import _  // error
         import    // error
