+++ tests/neg/parser-stability-23.scala
@@ -1,3 +1,3 @@
 object i0 {
-  import Ordering.{ implicitly as } (true: Boolean) match { case _: i1 as true } // error // error
+  import Ordering.{ implicitly as } (true: Boolean) match { case _: i1 as true } // error // error // error
 }
